requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.6
python-dotenv>=1.0.0
requests>=2.31
numpy>=1.24
scikit-learn>=1.3

pdfplumber>=0.10
PyPDF2>=3.0.0
scipy>=1.10
